pyproject.toml
[project]
name = "resume-analyser"
version = "0.1.0"
description = "AI agent to retrieve resume from gmail then get the github and make a github report"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
     "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.4",
    "aiosignal==1.3.2",
    "alembic==1.16.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appdirs==1.4.4",
    "asgiref==3.8.1",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "auth0-python==4.9.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "blinker==1.9.0",
    "build==1.2.2.post1",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.5.23",
    "click==8.2.1",
    "cohere==5.15.0",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "crewai==0.121.1",
    "crewai-tools==0.46.0",
    "cryptography==45.0.3",
    "dataclasses-json==0.6.7",
    "decorator==5.2.1",
    "deprecated==1.2.18",
    "deprecation==2.1.0",
    "distro==1.9.0",
    "docker==7.1.0",
    "docstring-parser==0.16",
    "durationpy==0.10",
    "embedchain==0.1.128",
    "et-xmlfile==2.0.0",
    "executing==2.2.0",
    "fastapi==0.115.9",
    "fastavro==1.11.1",
    "filelock==3.18.0",
    "flatbuffers==25.2.10",
    "frozenlist==1.6.0",
    "fsspec==2025.5.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.2",
    "google-api-python-client==2.170.0",
    "google-auth==2.40.2",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.2",
    "google-genai==1.17.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "gptcache==0.1.44",
    "greenlet==3.2.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.32.3",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "importlib-resources==6.5.2",
    "instructor==1.8.3",
    "ipython==9.2.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.8.2",
    "json-repair==0.46.0",
    "json5==0.12.0",
    "jsonpatch==1.33",
    "jsonpickle==4.1.0",
    "jsonpointer==3.0.0",
    "jsonref==1.1.0",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "kubernetes==32.0.1",
    "lancedb==0.22.1",
    "langchain==0.3.25",
    "langchain-cohere==0.3.5",
    "langchain-community==0.3.24",
    "langchain-core==0.3.63",
    "langchain-experimental==0.3.4",
    "langchain-openai==0.2.14",
    "langchain-text-splitters==0.3.8",
    "langsmith==0.3.43",
    "litellm==1.68.0",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mem0ai==0.1.102",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "networkx==3.5",
    "nodeenv==1.9.1",
    "numpy==2.2.6",
    "oauthlib==3.2.2",
    "onnxruntime==1.22.0",
    "openai==1.75.0",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.33.1",
    "opentelemetry-exporter-otlp-proto-common==1.33.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.33.1",
    "opentelemetry-exporter-otlp-proto-http==1.33.1",
    "opentelemetry-instrumentation==0.54b1",
    "opentelemetry-instrumentation-asgi==0.54b1",
    "opentelemetry-instrumentation-fastapi==0.54b1",
    "opentelemetry-proto==1.33.1",
    "opentelemetry-sdk==1.33.1",
    "opentelemetry-semantic-conventions==0.54b1",
    "opentelemetry-util-http==0.54b1",
    "orjson==3.10.18",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pdfminer-six==20250327",
    "pdfplumber==0.11.6",
    "pillow==11.2.1",
    "portalocker==2.10.1",
    "posthog==3.25.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "pure-eval==0.2.3",
    "pyarrow==20.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pygments==2.19.1",
    "pyjwt==2.10.1",
    "pymupdf==1.26.0",
    "pyparsing==3.2.3",
    "pypdf==5.5.0",
    "pypdf2==3.0.1",
    "pypdfium2==4.30.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "pyreadline3==3.5.4",
    "pyright==1.1.401",
    "pysbd==0.3.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytube==15.0.0",
    "pytz==2025.2",
    "pyvis==0.3.2",
    "pywin32==310",
    "pyyaml==6.0.2",
    "qdrant-client==1.14.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "schema==0.7.7",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "stack-data==0.6.3",
    "starlette==0.45.3",
    "sympy==1.14.0",
    "tabulate==0.9.0",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.20.3",
    "tomli==2.2.1",
    "tomli-w==1.2.0",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.16.0",
    "types-requests==2.32.0.20250515",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uv==0.7.8",
    "uvicorn==0.34.2",
    "watchfiles==1.0.5",
    "wcwidth==0.2.13",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "yarl==1.20.0",
    "zipp==3.22.0",
    "zstandard==0.23.0"
]
